IMR-72 refactor : move image to imgs
